Create/update miq template from local file
Previously we were making temp dirs and copying yaml template files
over if we found that the template object doesn't exist on the remote
host already.

Now we read in the template file locally using a `lookup('file', ...)`
and then use `oc_obj` to create/update the template object
automatically if they differ.